ctf_tools/requirements.txt
binwalk 
pillow
pyPdf2